+++ src/WarriorMonster.cpp
@@ -5,7 +5,7 @@
 
 WarriorMonster::WarriorMonster(
     QSvgRenderer* renderer,
-    short type,
+    size_t type,
     QGraphicsSvgItem* parent)
   : Monster(renderer, type, parent) {
   this->initializeMonster();
