+++ src/WarriorMonster.hpp
@@ -11,7 +11,7 @@ class WarriorMonster : public Monster {
 
   public: // Constructor and destructor
    explicit WarriorMonster(QSvgRenderer* renderer,
-                    short type = 0,
+                    size_t type = 0,
                     QGraphicsSvgItem* parent = nullptr);
    ~WarriorMonster() = default;
